fix: improve migrations and add dev docs (#35)
* fix: improve migrations and add dev docs

* prettier

+++ packages/server/src/migrations/replace_start_and_end_with_date_invoice.ts
@@ -2,6 +2,13 @@ import { Migration } from '@mikro-orm/migrations';
 
 export class MigrationReplaceStartAndEndWithDate extends Migration {
   async up(): Promise<void> {
+    if (
+      !(await this.ctx?.schema.hasColumn('invoice', 'start')) ||
+      !(await this.ctx?.schema.hasColumn('invoice', 'end'))
+    ) {
+      return;
+    }
+
     await this.ctx?.schema.alterTable('invoice', (table) => {
       table.dropColumn('start');
       table.renameColumn('end', 'date');
@@ -9,6 +16,10 @@ export class MigrationReplaceStartAndEndWithDate extends Migration {
   }
 
   async down(): Promise<void> {
+    if (!(await this.ctx?.schema.hasColumn('invoice', 'date'))) {
+      return;
+    }
+
     await this.ctx?.schema.alterTable('invoice', (table) => {
       table.date('start');
       table.renameColumn('date', 'end');
